[RFC] Merge White and BlackAndWhite themes
The app has a theme that is only light (White/Blank) and another that is
only dark (BlackAndWhite).
These themes are standalone and there is no way to use them following
the system theme.

Merge them as the light and dark variant, respectively, of the same
theme. This way the user can follow the system theme or simply leave
always the light or dark one (current behaviour).